+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Upcoming - 3.0 - WebAssembly
 ---
 - jsNet version 1.5 feature parity
+- Global bug fixes
+- Removed babel transpilation (it's 2018)
 
 # 2.1.0 - Optimizations
 ---